# API_DOCUMENTATION.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "Finance Data Processing API",
                 "description":  "Backend API for Finance Dashboard with Role-Based Access Control",
                 "contact":  {
                                 "name":  "Md Shahbaz Imam",
                                 "email":  "mdshahbazimam02@gmail.com"
                             },
                 "version":  "1.0.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Default Server"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
...
    "paths":  {
...
                  "/api/auth/register":  {
                                             "post":  {
                                                          "tags":  [
                                                                       "Authentication"
                                                                   ],
                                                          "summary":  "Register a new user",
                                                          "description":  "Creates a new user with VIEWER role by default",
                                                          "operationId":  "register",
                                                          "requestBody":  {
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/RegisterRequest"
                                                                                                                                  }
                                                                                                                   }
                                                                                          },
                                                                              "required":  true
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "OK",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/ApiResponseAuthResponse"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    }
                                                                        }
                                                      }
                                         },
                  "/api/auth/login":  {
                                          "post":  {
                                                       "tags":  [
                                                                    "Authentication"
                                                                ],
                                                       "summary":  "Login",
                                                       "description":  "Authenticate and receive a JWT token",
                                                       "operationId":  "login",
                                                       "requestBody":  {
                                                                           "content":  {
                                                                                           "application/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/LoginRequest"
                                                                                                                               }
                                                                                                                }
                                                                                       },
                                                                           "required":  true
                                                                       },
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "OK",
                                                                                     "content":  {
                                                                                                     "*/*":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/ApiResponseAuthResponse"
                                                                                                                            }
                                                                                                             }
                                                                                                 }
                                                                                 }
                                                                     }
                                                   }
                                      },
...
                  "/api/auth/me":  {
                                       "get":  {
                                                   "tags":  [
                                                                "Authentication"
                                                            ],
                                                   "summary":  "Get current user profile",
                                                   "description":  "Returns the authenticated user\u0027s profile info",
                                                   "operationId":  "getCurrentUser",
                                                   "responses":  {
                                                                     "200":  {
                                                                                 "description":  "OK",
                                                                                 "content":  {
                                                                                                 "*/*":  {
                                                                                                             "schema":  {
                                                                                                                            "$ref":  "#/components/schemas/ApiResponseUserResponse"
                                                                                                                        }
                                                                                                         }
                                                                                             }
                                                                             }
                                                                 },
                                                   "security":  [
                                                                    {
                                                                        "bearerAuth":  [

                                                                                       ]
                                                                    }
                                                                ]
                                               }
                                   }
              },
    "components":  {
                       "schemas":  {
...
                                       "RegisterRequest":  {
                                                               "required":  [
                                                                                "email",
                                                                                "name",
                                                                                "password"
                                                                            ],
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "name":  {
                                                                                               "maxLength":  50,
                                                                                               "minLength":  2,
                                                                                               "type":  "string"
                                                                                           },
                                                                                  "email":  {
                                                                                                "type":  "string"
                                                                                            },
                                                                                  "password":  {
                                                                                                   "maxLength":  2147483647,
                                                                                                   "minLength":  6,
                                                                                                   "type":  "string"
                                                                                               }
                                                                              }
                                                           },
                                       "ApiResponseAuthResponse":  {
                                                                       "type":  "object",
                                                                       "properties":  {
                                                                                          "success":  {
                                                                                                          "type":  "boolean"
                                                                                                      },
                                                                                          "message":  {
                                                                                                          "type":  "string"
                                                                                                      },
                                                                                          "data":  {
                                                                                                       "$ref":  "#/components/schemas/AuthResponse"
                                                                                                   },
                                                                                          "timestamp":  {
                                                                                                            "type":  "string",
                                                                                                            "format":  "date-time"
                                                                                                        }
                                                                                      }
                                                                   },
                                       "AuthResponse":  {
                                                            "type":  "object",
                                                            "properties":  {
                                                                               "token":  {
                                                                                             "type":  "string"
                                                                                         },
                                                                               "tokenType":  {
                                                                                                 "type":  "string"
                                                                                             },
                                                                               "user":  {
                                                                                            "$ref":  "#/components/schemas/UserResponse"
                                                                                        }
                                                                           }
                                                        },
                                       "UserResponse":  {
                                                            "type":  "object",
                                                            "properties":  {
                                                                               "id":  {
                                                                                          "type":  "integer",
                                                                                          "format":  "int64"
                                                                                      },
                                                                               "name":  {
                                                                                            "type":  "string"
                                                                                        },
                                                                               "email":  {
                                                                                             "type":  "string"
                                                                                         },
                                                                               "role":  {
                                                                                            "type":  "string",
                                                                                            "enum":  [
                                                                                                         "VIEWER",
                                                                                                         "ANALYST",
                                                                                                         "ADMIN"
                                                                                                     ]
                                                                                        },
                                                                               "status":  {
                                                                                              "type":  "string",
                                                                                              "enum":  [
                                                                                                           "ACTIVE",
                                                                                                           "INACTIVE"
                                                                                                       ]
                                                                                          },
                                                                               "createdAt":  {
                                                                                                 "type":  "string",
                                                                                                 "format":  "date-time"
                                                                                             }
                                                                           }
                                                        },
                                       "LoginRequest":  {
                                                            "required":  [
                                                                             "email",
                                                                             "password"
                                                                         ],
                                                            "type":  "object",
                                                            "properties":  {
                                                                               "email":  {
                                                                                             "type":  "string"
                                                                                         },
                                                                               "password":  {
                                                                                                "type":  "string"
                                                                                            }
                                                                           }
                                                        },
...
                                       "ApiResponseUserResponse":  {
                                                                       "type":  "object",
                                                                       "properties":  {
                                                                                          "success":  {
                                                                                                          "type":  "boolean"
                                                                                                      },
                                                                                          "message":  {
                                                                                                          "type":  "string"
                                                                                                      },
                                                                                          "data":  {
                                                                                                       "$ref":  "#/components/schemas/UserResponse"
                                                                                                   },
                                                                                          "timestamp":  {
                                                                                                            "type":  "string",
                                                                                                            "format":  "date-time"
                                                                                                        }
                                                                                      }
                                                                   },
...
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "description":  "JWT Bearer Token. Format: Bearer \u003ctoken\u003e",
                                                                  "in":  "header",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}