# src/main/java/com/financeapp/dto/response/AuthResponse.java
package com.financeapp.dto.response;

import lombok.Builder;
import lombok.Data;

@Data
@Builder
public class AuthResponse {
    private String token;
    private String tokenType;
    private UserResponse user;
}