# src/main/java/com/financeapp/dto/response/UserResponse.java
package com.financeapp.dto.response;

import com.financeapp.enums.Role;
import com.financeapp.enums.UserStatus;
import lombok.Builder;
import lombok.Data;

import java.time.LocalDateTime;

@Data
@Builder
public class UserResponse {
    private Long id;
    private String name;
    private String email;
    private Role role;
    private UserStatus status;
    private LocalDateTime createdAt;
}